requirements.txt
azure-identity
#azure-core
#azure-common
#azure-mgmt-core
#azure-mgmt-resource
azure-mgmt-containerservice
